Making a more compact code...
As you can see in index.php, next code will be a lot more compact.
A lot.
I still have to finish developing this and the new style, but after
finishing that I think I'm going to delete almost all the files, with an
exception for the font file, the stylev2 file, index.php, footer, form
and README.